pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "erenshor-wiki"
version = "0.1.0"
description = "Generate MediaWiki content and icons from Erenshor game database"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "Erenshor Wiki Team"},
]
keywords = ["erenshor", "wiki", "mediawiki", "game-data"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Games/Entertainment",
    "Topic :: Text Processing :: Markup",
]

dependencies = [
    "pillow>=11.2.1",
    "pydantic-settings>=2.10.1",
    "httpx>=0.27.0",
    "sqlmodel>=0.0.24",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "dependency-injector>=4.48.1",
    "structlog>=24.1.0",
    "jinja2>=3.1.6",
    "sqlalchemy>=2.0.41",
    "mwparserfromhell>=0.6.6",
    "platformdirs>=4.0.0",
    "google-auth>=2.0.0",
    "google-api-python-client>=2.0.0",
]

[project.urls]
Homepage = "https://erenshor.wiki.gg"
Repository = "https://github.com/glockyco/erenshor-wiki"

[project.scripts]
erenshor-wiki = "erenshor.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/erenshor"]

[dependency-groups]
dev = [
    "gitlint>=0.19.1",
    "mypy>=1.17.1",
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.12.11",
    "ty>=0.0.1a20",
]

[tool.ruff]
target-version = "py311"
src = ["src"]

[tool.ruff.lint]

[tool.ruff.format]

[tool.ty]

[tool.mypy]
strict = true
mypy_path = "src"
explicit_package_bases = true
namespace_packages = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=src/erenshor",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-branch",
]
markers = [
    "integration: integration tests with real database and file I/O",
    "slow: tests that take more than a few seconds",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src/erenshor"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
